\name{FastMath-package}
\alias{FastMath-package}
\alias{FastMath}
\docType{package}
\title{
\packageTitle{FastMath}
}
\description{
\packageDescription{FastMath}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FastMath}
\packageIndices{FastMath}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FastMath}

Maintainer: \packageMaintainer{FastMath}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
