% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinv.R
\name{pinv}
\alias{pinv}
\title{Matrix pseudoinverse}
\arguments{
\item{V}{A numeric matrix}

\item{tol}{A relative tolerance for detecting zero singular values}
}
\value{
The Moore-Penrose pseudoinverse for \code{V}.
}
\description{
Compute the Moore-Penrose pseudoinverse using the SVD approximation.
}
\details{
To use the same tolerance as \code{MASS::ginv}, set the tolerance parameter 
to \code{sqrt(.Machine$double.eps)}.
}
